+++ pyproject.toml
@@ -52,6 +52,9 @@ packages = [
   "zapzap.core.platform",
   "zapzap.core.theme",
   "zapzap.features",
+  "zapzap.features.accounts",
+  "zapzap.features.accounts.domain",
+  "zapzap.features.accounts.models",
   "zapzap.features.browser",
   "zapzap.features.browser.controllers",
   "zapzap.features.browser.ui",
@@ -79,7 +82,6 @@ packages = [
   "zapzap.features.permissions",
   "zapzap.features.startup",
   "zapzap.features.tray",
-  "zapzap.models",
   "zapzap.assets",
   "zapzap.assets.icons",
   "zapzap.assets.themes",
