+++ zapzap/features/accounts/__init__.py
@@ -1,0 +1,1 @@
+"""Accounts feature."""
