+++ zapzap/features/accounts/domain/__init__.py
@@ -1,0 +1,1 @@
+"""Account domain models."""
