+++ zapzap/features/accounts/models/__init__.py
@@ -1,0 +1,1 @@
+"""Account presentation models."""
